# fisherfans-openapi-min.yaml
openapi: 3.1.1
info:
  title: FisherFans API
  version: "3"
  description: API for FisherFans
servers:
  - url: http://127.0.0.1:8000/api/
...
paths:
...
  /v1/users/{userId}/boats:
    parameters:
      - $ref: '#/components/parameters/PathUserId'
    get:
      tags: [Users]
      summary: Retrieve the list of boats of an user
      operationId: listUserBoats
      responses:
        '200':
          $ref: '#/components/responses/ListBoats'
        '400':
          description: The request is malformed or invalid.
        '401':
          description: The client is not authenticated.
        '403':
          description: The client is authenticated but lacks access rights.
        '404':
          description: The requested resource does not exist.
        '500':
          description: Unexpected internal server error.
  /v1/users/{userId}/trips:
    parameters:
      - $ref: '#/components/parameters/PathUserId'
    get:
      tags: [Users]
      summary: Retrieve the list of trips of an user
      operationId: listUserTrips
      responses:
        '200':
          description: Response containing a list of trips of an user.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Trip'
        '400':
          description: The request is malformed or invalid.
        '401':
          description: The client is not authenticated.
        '403':
          description: The client is authenticated but lacks access rights.
        '404':
          description: The requested resource does not exist.
        '500':
          description: Unexpected internal server error.
  /v1/users/{userId}/bookings:
    parameters:
      - $ref: '#/components/parameters/PathUserId'
    get:
      tags: [Users]
      summary: Retrieve the list of bookings of an user
      operationId: listUserBookings
      responses:
        '200':
          description: Response containing a list of bookings of an user.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
        '400':
          description: The request is malformed or invalid.
        '401':
          description: The client is not authenticated.
        '403':
          description: The client is authenticated but lacks access rights.
        '404':
          description: The requested resource does not exist.
        '500':
          description: Unexpected internal server error.

  # --------------------
  # RESSOURCE CARNET DE PÊCHE (Logbook Pages)
  # --------------------
  # Les pages du carnet sont généralement imbriquées sous l'utilisateur
...
components:
  schemas:
...
    Booking:
      type: object
      properties:
        reservationId:
          type: integer

        date:
          type: string
          format: date

        numberOfSeats:
          type: integer

        totalPrice:
          type: number
          format: float

        user_id:
          type: integer

        trip_id:
          type: integer
    Trip:
      type: object
      properties:
        tripId:
          type: integer

        title:
          type: string

        practicalInformation:
          type: string

        tripType:
          type: string
          enum:
            - JOURNALIERE
            - RECURRENTE

        priceType:
          type: string
          enum:
            - GLOBAL
            - PAR_PERSONNE

        startDates:
          type: array
          items:
            type: string
            format: date

        endDates:
          type: array
          items:
            type: string
            format: date

        departureTimes:
          type: array
          items:
            type: string
            format: time

        endTimes:
          type: array
          items:
            type: string
            format: time

        passengerCount:
          type: integer

        user_id:
          type: integer

        boat_id:
          type: integer

        price:
          type: number
          format: float
    Boat:
      type: object
      properties:
        boatId:
          type: integer

        name:
          type: string

        description:
          type: string

        brand:
          type: string

        manufacturingYear:
          type: integer

        photoUrl:
          type: string

        permitType:
          type: string
          enum:
            - COTIER
            - FLUVIAL

        boatType:
          type: string
          enum:
            - OPEN
            - CABINE
            - CATAMARAN
            - VOILIER
            - JETSKI
            - CANOE

        equipment:
          type: array
          uniqueItems: true
          items:
            type: string
            enum:
              - SONDEUR
              - VIVIER
              - ECHELLE
              - GPS
              - PORTECANNES
              - RADIO_VHF

        depositAmount:
          type: number
          format: float

        maxCapacity:
          type: integer

        numberOfBeds:
          type: integer

        homePort:
          type: string

        latitude:
          type: number
          format: float

        longitude:
          type: number
          format: float

        engineType:
          type: string
          enum:
            - DIESEL
            - ESSENCE
            - AUCUN

        enginePower:
          type: number
          format: float

        user_id:
          type: integer
...
  parameters:
...
    PathUserId:
      name: userId
      in: path
      schema: {type: integer, format: int64}
      description: Unique user ID.
...
  responses:
...
    ListBoats:
      description: Response containing a list of boats of an user.
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Boat'    
  securitySchemes:
    security-model1:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - security-model1: []